Add the `-o/--output` options.
Because we can compute several URL (from STDIN), the reporter may output
their results in different files. The `-o/--output` option takes a
valid filename as value. The given filename will receive the output from
the reporter (only the output, not the error). This is also useful when
dealing with a single URL (it can replace the `>` Shell redirection).

This option understands a special keyword: `%HASH%`, which represents
the SHA1 of the currently computed URL. So for instance:

    $ echo file:///a.html\nfile:///b.html | \
          pa11y -0 -r html -o 'output/%HASH%.html'

will produce 2 HTML reports in the `output/` directory.

To produce an index, simply insert a command between the pipe for
instance.