Adding an option for running JS before testing
This is an attempt to add the new functionality suggested in issue #98.

This will add the `--injectJs` option to pass in a file path to a
JavaScript file to be run pa11y is run.

That JavaScript file has to `console.log('pa11y: script complete')`
before pa11y will be run.

I am not 100% sure how I feel about this `console.log` approach but it
felt like it kept the supplied JS simple and without any phantomjs
knowledge required compared to the callback approach.

Tests still need to be written.