Adds the ability to hide elements from pa11y testing
This will allow areas of the page to be hidden before running HTML code
sniffer against the page. This relies on the fact that HTML code sniffer
will ignore anything thats `visibility` style is set to `hidden`.

The option accepts a string which allows for coma seperated selectors in
the same way as `querySelectorAll`.

e.g. "#modal, .adverts, div[data-hide]"