Update all of the dependencies
Now using caret versions for everything. This is generally safe to do,
and all of the dependencies we use are mature enough to not be
introducing breaking changes under minor/patch releases.

Also I've made sure that https://nightli.es/ runs for all of the Pa11y
projects so that we're alerted early if a dependency ever does start to
fail the build.