Add phantom exit code 127 to troubleshooting guide
As we've had two issues already mentioning this error (#175 and #270) I thought that it would be a good idea to add it to the troubleshooting guide as there's not much that we can do about problems with phantomjs.

I'm not extremely happy with the wording, so feel free to suggest any improvements.