Allow sharing a Chrome browser instance
This PR adds the ability to pass in a Puppeteer Browser instance to be
shared across multiple Pa11y test runs, this should be less resource
intensive and we need this work in place for Pa11y CI and probably
Sidekick.

I've added some warnings to the README for anyone using this themselves,
but mostly I think it'll be used by our own derivative projects.

Resolves #298