Supports aXe test runner (#408)
* Support aXe test runner (and other custom runners)

We now support both the aXe and HTML CodeSniffer test runners - You can
run Pa11y using one or both of these. I've done this by untangling HTML
CodeSniffer from the core code and writing two new modules:

  - https://github.com/pa11y/pa11y-runner-axe
  - https://github.com/pa11y/pa11y-runner-htmlcs

These come bundled with Pa11y so there's no need to install separately
for now. They'll be considered stable when we're ready to release the
code on this branch.

* Fix a copy/paste error

* Action review comments from @andrewmee

* Update to the latest version of the aXe runner

This fixes the issue with The Verge

* Update documentation for HTMLCS-only configurations

* Use stable versions of HTMLCS and aXe test runners

* Tests error is thrown when ignoreUrl is used without a page

* Sync package-lock.json to package.json

* Caches promises for runner scripts rather than global state