Downgrades nyc to 13.2.0 (#484)
There's a change in versions greater than 13.2.0 where nyc fails to include some of the files in the report. There's probably a config somewhere that can address this, I can't seem to be able to find it so for now I'm just downgrading nyc so the coverage works again as expected.