Extract option functions to a new separate module
Extracts the functions in `lib/pa11y` related to argument parsing and option processing (`parseArguments`, `verifyOptions`, `defaultOptions` and `sanitizeUrl`) to a separate module.

Only `parseArguments` and `verifyOptions` function is used somewhere else, so the module has a clean, simple interface.

No changes to the existing tests were necessary in order for them to pass, but I've extracted the tests related to these options to their own file.