Replace Travis with Github Actions for v5 (#554)
* Replace Travis with Github Actions for v5
* Increase integration tests timeout from 10 to 120s
* Add (longer) timeout to `lib/pa11y` test setup
* Remove nyc from project

The project has been consistently failing to run the integrations tests
on Github Actions, with tests randomly timing out.

After a lot of investigation and debugging I get the feeling that this
may be a side-effect of using nyc. I think that maybe the
instrumentation that nyc is adding to the tests is what is causing them
to be very slow.

We're using an old version of `nyc` but, as the project has been without
a working CI/CD for so long, I think I rather remove nyc from now and
hope that this will be enough to get the tests running again. We can
have a look at adding nyc back at a later date.