Make the console logging code more resilient
Currently the code used for showing the browsers console output is being run even when debugging is enabled. This is causing pa11y to crash when puppeteer returns an object with circular references.

This formats the output in a different way so it will display `source:line:column message` for every line of console output and avoid the use of puppeteer's [`consoleMessage.args()`][1] which was causing the circular reference in the first place.
[1]: https://pptr.dev/#?product=Puppeteer&version=v9.1.1&show=api-consolemessageargs

We could also prevent the event listeners from being set if debugging is enabled but, as that would required further changes, I'll leave it for now so we can release a quick bugfix.

Fixes #634.
Closes #635.